+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+
+## [1.2.0] (2023-05-30)
+
  * Add EXIF, IPTC, XMP metadata handling for JPEG, PNG, WEBP, GIF, HEIC, JXL, AVIF. [#93], [#95], [#96]
  * Improve file name hashing algorithm. [#90]
  * Fix wrong exception class being thrown.
@@ -135,7 +138,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
  * Initial release
 
-[Unreleased]: https://github.com/contao/image/compare/1.1.2...HEAD
+[Unreleased]: https://github.com/contao/image/compare/1.2.0...1.x
+[1.2.0]: https://github.com/contao/image/compare/1.1.2...1.2.0
 [1.1.2]: https://github.com/contao/image/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/contao/image/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/contao/image/compare/1.0.3...1.1.0